+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.10
 
+* 修正事物类命名 [cn.simplifydb.database.TransactionException]
 * 新增可配置自动全局还原html 实体 cn.simplifydb.database.config.ConfigProperties.UNESCAPE_HTML
 
 -------------------------------------------------------------------------------------------------------------